+++ rust/services/call/engine/src/evm/env.rs
@@ -1,45 +1,25 @@
-use std::{
-    collections::HashMap,
-    fmt::{self, Debug, Formatter},
-};
+use std::{collections::HashMap, fmt::Debug};
 
 use alloy_primitives::{BlockHash, BlockNumber, ChainId};
 use block_header::EvmBlockHeader;
+use call_common::RevmDB;
 use chain::{ChainSpec, ForkError};
 use derive_more::{Deref, DerefMut, From, Into, IntoIterator};
-use revm::{
-    primitives::{CfgEnvWithHandlerCfg, HandlerCfg, SpecId},
-    DatabaseRef,
-};
+use revm::primitives::{CfgEnvWithHandlerCfg, HandlerCfg, SpecId};
 
 pub mod cached;
 pub mod factory;
 pub mod location;
 
 /// The environment to execute the contract calls in.
-pub struct EvmEnv<D: DatabaseRef + Send + Sync> {
+#[derive(Debug)]
+pub struct EvmEnv<D: RevmDB> {
     pub db: D,
     pub cfg_env: CfgEnvWithHandlerCfg,
     pub header: Box<dyn EvmBlockHeader>,
 }
 
-impl<D> Debug for EvmEnv<D>
-where
-    D: DatabaseRef + Send + Sync + Debug,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        f.debug_struct("EvmEnv")
-            .field("db", &self.db)
-            .field("cfg_env", &self.cfg_env)
-            .field("header", &self.header)
-            .finish()
-    }
-}
-
-impl<D> EvmEnv<D>
-where
-    D: DatabaseRef + Send + Sync,
-{
+impl<D: RevmDB> EvmEnv<D> {
     /// Creates a new environment.
     /// It uses the default configuration for the latest specification.
     pub fn new(db: D, header: Box<dyn EvmBlockHeader>) -> Self {