+++ rust/guest_wrapper/artifacts/chain_guest/CHANGELOG.md
@@ -2,3 +2,4 @@
   * `9471af8ec35f9f6ffa891b504cc73735ebdc7488b613cdb0cf5cfb16d2b4627a` – Initial version deployed in December 2024
   * `18eac1fe344b5f8ce19ad5695ea43b5fe85dafb36c6158d51331adb4ea33eb88` – Teleport is a no-op on local-testnet (anvil) [#1721](https://github.com/vlayer-xyz/vlayer/pull/1721)
   * `907fa4c2e540a151e2053e66f638cb65932af0c9815324f16481d1922b4164aa` – History cleanup, seed with first deployed ELF ID [#1722](https://github.com/vlayer-xyz/vlayer/pull/1722)
+  * `e1afa09f684c5a099b78b0a1425a7d9bec3ec8bd4d7a121f29a62f3bb0172573` – Teleport logging